Modified the UDP_Sock class for java to relate closer to the standard of other Comm classes
UDP_Sock startReceive now takes a "Handable" object as an argument. This
replaces having to creating a class from UDP_Sock to create a handler.

UDP_Sock is now standardized to an interface (NetworkInterface)

The interface Handable was added

The test programs have been updated to use the new Handable interface,
but have not yet been tested.